+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Old Details block that requires ACF
 - `aria-labelledby` attribute from `<div>` element with no specified role.
 
+### Changed
+
+- Category title and slug updated from 'Custom' / 'govuk-custom' to 'GOV.UK Design System' / 'govuk-components'
+
 ## [0.4.3] - 2024-09-09
 
 ### Fixed